+++ NEWS.md
@@ -1,3 +1,5 @@
+# timetk (development version)
+
 # timetk 2.6.0
 
 __New Functions__