+++ NEWS.md
@@ -6,11 +6,14 @@ __Improvements__
 
 __Fixes__
 
-Reduce "New names" messages. 
+- Reduce "New names" messages. 
 ```
 #> New names: 
 #> * NA -> ...1 
 ```
+
+- Remove dependency on `lazyeval`. #24
+- Fix deprecated functions:  `select_()` used with `tk_xts_()`. #52
 
 # timetk 2.6.0